+++ src/AST/ASTLocalVar.cpp
@@ -52,7 +52,7 @@ std::string ASTLocalVar::str() const {
             Super(ASTExprStmt::str()).
             Attr("Type", Type).
             Attr("Name", Name).
-            Attr("VarKind", (int) VarKind).
+            Attr("VarKind", (uint64_t) VarKind).
             Attr("Constant", Constant).
             End();
 }
